Add median and median_approx for Spark engine
