+++ core/src/main/java/io/ddf/etl/ASqlHandler.java
@@ -36,8 +36,8 @@ public abstract class ASqlHandler extends ADDFFunctionalGroupHandler implements 
   }
 
   /**
-   * @brief Show tables in the database.
    * @return The table names.
+   * @brief Show tables in the database.
    */
   public SqlResult showTables() throws DDFException {
     List<String> tableNames = new ArrayList<String>();
@@ -53,27 +53,27 @@ public abstract class ASqlHandler extends ADDFFunctionalGroupHandler implements 
   }
 
   /**
-   * @brief Get the column information of this table.
-   * @param name The URI or the name of the ddf.
+   * @param name      The URI or the name of the ddf.
    * @param namespace The namespace.
    * @return The column information.
    * @throws DDFException
+   * @brief Get the column information of this table.
    */
   private SqlResult describeTable(String name, String namespace)
           throws DDFException {
     DDF ddf = null;
     try {
-        ddf = this.getManager().getOrRestoreDDFUri(name);
+      ddf = this.getManager().getOrRestoreDDFUri(name);
     } catch (Exception e) {
-        if (null == namespace) {
-            throw new DDFException("ERROR: there is no ddf " + name);
-        }
-        try {
-            mLog.info("trying to restore " + "ddf://" + namespace + "/" + name);
-            ddf = this.getManager().getOrRestoreDDFUri("ddf://" + namespace + "/" + name);
-        } catch (Exception e2) {
-            throw new DDFException("ERROR: there is no ddf " + name);
-        }
+      if (null == namespace) {
+        throw new DDFException("ERROR: there is no ddf " + name);
+      }
+      try {
+        mLog.info("trying to restore " + "ddf://" + namespace + "/" + name);
+        ddf = this.getManager().getOrRestoreDDFUri("ddf://" + namespace + "/" + name);
+      } catch (Exception e2) {
+        throw new DDFException("ERROR: there is no ddf " + name);
+      }
     }
 
     int colSize = ddf.getNumColumns();
@@ -94,9 +94,9 @@ public abstract class ASqlHandler extends ADDFFunctionalGroupHandler implements 
                              Integer maxRows,
                              DataSourceDescriptor dataSource) throws DDFException {
     return this.sqlHandle(command,
-                          maxRows,
-                          dataSource,
-                          new TableNameReplacer(this.getManager(), dataSource));
+            maxRows,
+            dataSource,
+            new TableNameReplacer(this.getManager(), dataSource));
   }
 
 
@@ -107,24 +107,32 @@ public abstract class ASqlHandler extends ADDFFunctionalGroupHandler implements 
     // If the user specifies the datasource, we should directly send the sql
     // command to the sql engine.
     if (dataSource != null) {
-        // TODO: add support for other datasource.
-        if (dataSource instanceof JDBCDataSourceDescriptor) {
-            // It's the jdbc datasource.
+      // TODO: add support for other datasource.
+      if (dataSource instanceof JDBCDataSourceDescriptor) {
+        // It's the jdbc datasource.
+        return this.sql(sqlcmd, maxRows, dataSource);
+      }
+      SQLDataSourceDescriptor sqlDataSourceDescriptor = (SQLDataSourceDescriptor) dataSource;
+      if (sqlDataSourceDescriptor == null) {
+        throw new DDFException("ERROR: Handling datasource");
+      }
+      if (sqlDataSourceDescriptor.getDataSource() != null) {
+        switch (sqlDataSourceDescriptor.getDataSource()) {
+          case "SparkSQL":
+          case "spark":
+          case "Spark":
+            String sql = sqlcmd;
+
+            if (sql.toLowerCase().trim().startsWith("select")) {
+              sql = handleSqlExpression(sql);
+            }
+
+            return this.sql(sql, maxRows, dataSource);
+          default:
+            //throw new DDFException("ERROR: Unrecognized datasource");
             return this.sql(sqlcmd, maxRows, dataSource);
         }
-        SQLDataSourceDescriptor sqlDataSourceDescriptor = (SQLDataSourceDescriptor)dataSource;
-        if (sqlDataSourceDescriptor == null) {
-            throw  new DDFException("ERROR: Handling datasource");
-        }
-        if (sqlDataSourceDescriptor.getDataSource() != null) {
-            switch (sqlDataSourceDescriptor.getDataSource()) {
-                case "SparkSQL":case "spark":case "Spark":
-                    return this.sql(sqlcmd, maxRows, dataSource);
-                default:
-                    //throw new DDFException("ERROR: Unrecognized datasource");
-                    return this.sql(sqlcmd, maxRows, dataSource);
-            }
-        }
+      }
     }
 
 
@@ -135,28 +143,29 @@ public abstract class ASqlHandler extends ADDFFunctionalGroupHandler implements 
       Statement statement = parserManager.parse(reader);
       if (statement instanceof ShowTables) {
         return this.showTables();
-      } else if (statement instanceof  DescribeTable){
-        return this.describeTable(((DescribeTable)statement).getName()
+      } else if (statement instanceof DescribeTable) {
+        return this.describeTable(((DescribeTable) statement).getName()
                 .getName(), tableNameReplacer.getNamespace());
-      } else if (statement instanceof  Select) {
+      } else if (statement instanceof Select) {
         // Standard SQL.
-          statement = tableNameReplacer.run(statement);
-          this.mLog.info("Reformulate SQL to " + statement.toString());
-          return this.sql(statement.toString(), maxRows, dataSource);
+        statement = tableNameReplacer.run(statement);
+        this.mLog.info("Reformulate SQL to " + statement.toString());
+        String sql = handleSqlExpression(statement.toString());
+        return this.sql(sql, maxRows, dataSource);
       } else if (statement instanceof Drop) {
-          // TODO: +rename
-          return null;
+        // TODO: +rename
+        return null;
       } else {
-          throw  new DDFException("ERROR: Only show tables, describe tables, " +
-                  "select, drop, and rename operations are allowed on ddf");
+        throw new DDFException("ERROR: Only show tables, describe tables, " +
+                "select, drop, and rename operations are allowed on ddf");
       }
     } catch (JSQLParserException e) {
-        throw  new DDFException(" SQL Syntax ERROR: " + e.getCause().getMessage
-                ().split("\n")[0]);
+      throw new DDFException(" SQL Syntax ERROR: " + e.getCause().getMessage
+              ().split("\n")[0]);
     } catch (DDFException e) {
-        throw e;
+      throw e;
     } catch (Exception e) {
-        throw new DDFException(e);
+      throw new DDFException(e);
     }
   }
 
@@ -166,11 +175,12 @@ public abstract class ASqlHandler extends ADDFFunctionalGroupHandler implements 
                            DataSourceDescriptor dataSource,
                            DataFormat dataFormat) throws DDFException {
     return sql2ddfHandle(command,
-                         schema,
-                         dataSource,
-                         dataFormat,
-                         new TableNameReplacer(this.getManager(), dataSource));
+            schema,
+            dataSource,
+            dataFormat,
+            new TableNameReplacer(this.getManager(), dataSource));
   }
+
   public DDF sql2ddfHandle(String command,
                            Schema schema,
                            DataSourceDescriptor dataSource,
@@ -178,23 +188,29 @@ public abstract class ASqlHandler extends ADDFFunctionalGroupHandler implements 
                            TableNameReplacer tableNameReplacer) throws DDFException {
 
     if (dataSource != null) {
-        if (dataSource instanceof JDBCDataSourceDescriptor) {
+      if (dataSource instanceof JDBCDataSourceDescriptor) {
+        return this.sql2ddf(command, schema, dataSource, dataFormat);
+      }
+      SQLDataSourceDescriptor sqlDataSourceDescriptor = (SQLDataSourceDescriptor) dataSource;
+      if (sqlDataSourceDescriptor == null) {
+        throw new DDFException("ERROR: Handling datasource");
+      }
+      if (sqlDataSourceDescriptor.getDataSource() != null) {
+        switch (sqlDataSourceDescriptor.getDataSource()) {
+          case "SparkSQL":
+          case "spark":
+          case "Spark":
+            String sql = command;
+            if (sql.toLowerCase().trim().startsWith("select")) {
+              sql = handleSqlExpression(sql);
+            }
+            return this.sql2ddf(sql, schema, dataSource, dataFormat);
+          default:
+            // throw new DDFException("ERROR: Unrecognized datasource:
+            // " + dataSource);
             return this.sql2ddf(command, schema, dataSource, dataFormat);
         }
-        SQLDataSourceDescriptor sqlDataSourceDescriptor = (SQLDataSourceDescriptor)dataSource;
-        if (sqlDataSourceDescriptor == null) {
-            throw  new DDFException("ERROR: Handling datasource");
-        }
-        if (sqlDataSourceDescriptor.getDataSource() != null) {
-            switch (sqlDataSourceDescriptor.getDataSource()) {
-                case "SparkSQL":case "spark":case "Spark":
-                    return this.sql2ddf(command, schema, dataSource, dataFormat);
-                default:
-                    // throw new DDFException("ERROR: Unrecognized datasource:
-                    // " + dataSource);
-                    return this.sql2ddf(command, schema, dataSource, dataFormat);
-            }
-        }
+      }
     }
 
     this.mLog.info("Handle SQL: " + command);
@@ -203,21 +219,35 @@ public abstract class ASqlHandler extends ADDFFunctionalGroupHandler implements 
     try {
       Statement statement = parserManager.parse(reader);
       if (!(statement instanceof Select)) {
-        throw  new DDFException("ERROR: Only select is allowed in this sql2ddf");
+        throw new DDFException("ERROR: Only select is allowed in this sql2ddf");
       } else {
         statement = tableNameReplacer.run(statement);
         this.mLog.info("Reformulate SQL to " + statement.toString());
+        String sql = handleSqlExpression(statement.toString());
+
         // TODO(fanj) optimization here;
-        return this.sql2ddf(statement.toString(), schema, dataSource,
-                      dataFormat);
+        return this.sql2ddf(sql, schema, dataSource,
+                dataFormat);
 
       }
     } catch (JSQLParserException e) {
-        throw  new DDFException(" SQL Syntax ERROR: " + e.getCause().getMessage().split("\n")[0]);
+      throw new DDFException(" SQL Syntax ERROR: " + e.getCause().getMessage().split("\n")[0]);
     } catch (DDFException e) {
-        throw e;
+      throw e;
     } catch (Exception e) {
-        throw new DDFException(e);
+      throw new DDFException(e);
     }
+  }
+
+  /**
+   * This is a util function that should be implemented by sub-classes to handle engine-specific expression
+   * Use-cases:
+   * - Support a new UDF that is mapped into existing one. For example, median() which is mapped to percentile()
+   *
+   * @param sql The SQL query to handle
+   * @return modified sql
+   */
+  protected String handleSqlExpression(String sql) throws DDFException {
+    return sql;
   }
 }
